+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
--
+
+### Fixed
+- Show Warning if GPX file does not have time stamps
+
+### Added
+
 
 ## [1.2] - 2023-05-01